# Conflicts:
#	FertiFront/src/app/servicios/analisis.service.ts
#	FertiWeb/src/main/resources/dataH2.sql
#	FertiWeb/src/main/resources/dataMySql.sql

+++ FertiWeb/src/main/java/controladores/ControladorDatosParcela.java
@@ -6,11 +6,20 @@ import java.util.stream.Collectors;
 
 import org.dozer.DozerBeanMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.transaction.annotation.Transactional;
 
+import constantes.MensajesConstantes;
 import dominio.Parcela;
 import dto.DTOParcela;
+import excepciones.ExcepcionClaveForanea;
+import persistencia.entidad.AnalisisEntidad;
+import dto.DTOParcelaPaginado;
+import dto.Paginador;
 import persistencia.entidad.ParcelaEntidad;
+import persistencia.repositorio.AnalisisRepository;
 import persistencia.repositorio.ParcelaRepository;
 
 public class ControladorDatosParcela extends ControladorDatos {
@@ -20,6 +29,9 @@ public class ControladorDatosParcela extends ControladorDatos {
 
 	@Autowired
 	private ParcelaRepository parcelaRepository;
+
+	@Autowired
+	private AnalisisRepository analisisRepository;
 
 	@Autowired
 	private ControladorDatosLugar controladorDatosLugar;
@@ -100,12 +112,37 @@ public class ControladorDatosParcela extends ControladorDatos {
 
 	@Transactional
 	public void eliminarParcela(Long codigoParcela) {
+		List<AnalisisEntidad> analisis = analisisRepository.findByCodigoParcela(codigoParcela);
+		if (!analisis.isEmpty())
+			throw new ExcepcionClaveForanea(MensajesConstantes.ERROR_ASOCIACION_ANALISIS);
 		parcelaRepository.deleteById(codigoParcela);
 
 	}
 
 	public Object consultarParcelaPorUsuario(Long cedula) {
 		List<Parcela> listaParcela = mapearListaADominio(parcelaRepository.findAllByCodigoUsuario(cedula));
-		return construirListaDTO(listaParcela); 
+		return construirListaDTO(listaParcela);
+	}
+
+	public DTOParcelaPaginado consultarTodosPaginado(int pagina) {
+		Pageable paginador = PageRequest.of(pagina, 10);
+		return construirDTOPaginado(parcelaRepository.findAll(paginador));
+	}
+
+	private DTOParcelaPaginado construirDTOPaginado(Page<ParcelaEntidad> resultado) {
+		Paginador paginador = new Paginador();
+		paginador.setNumeroPagina(resultado.getNumber());
+		paginador.setTamano(resultado.getNumberOfElements());
+		paginador.setTotalElementos(resultado.getTotalElements());
+		paginador.setTotalPaginas(resultado.getTotalPages());
+		DTOParcelaPaginado dtoParcelaPaginado = new DTOParcelaPaginado();
+		dtoParcelaPaginado.setParcela(construirListaDTO(mapearListaADominio(resultado.getContent())));
+		dtoParcelaPaginado.setPaginador(paginador);
+		return dtoParcelaPaginado;
+	}
+
+	public DTOParcelaPaginado consultarConFiltroUsuario(String filtro, int pagina) {
+		Pageable paginador = PageRequest.of(pagina, 10);
+		return construirDTOPaginado(parcelaRepository.findAllLikeUsuarioNombre(filtro, paginador));
 	}
 }
